+++ src/Reseau.jl
@@ -14,7 +14,6 @@ export TCP, TLS
 
 include("0_memory_compat.jl")
 include("0_gcsafe_ccall_compat.jl")
-include("1_eventloops.jl")
 include("2_socket_ops.jl")
 include("3_iopoll.jl")
 include("4_tcp.jl")
